Allows user to pass versioned awscli zip bundles
Allows user to pass versioned awscli zip bundles

Before, the filename for the src zip bundle was hard-coded to
"awscli-bundle.zip". This meant that users could not use the
versioned bundles because there was no way to pass in the url
to the bundle file. AWS hosts versioned bundles using the form
"awscli-bundle-<version>.zip".

This patch allows users to pass the full URL to the zip bundle,
rather than just the host/path. It continues to default to the
un-versioned (latest) bundle.